pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pseudonym"
version = "0.7.0"
authors = [
  { name="Computerome", email="computerome@dtu.dk" },
]
description = "Computerome Pseudonymization"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
  'PyCryptodome',
  'Marshmallow',
  'dataclasses; python_version < "3.7"'
]

[project.urls]
"Homepage" = "https://gitlab.computerome.dk/delphi/pseudonymization"
